requirements.txt
Django==5.2
django-ckeditor==6.7.2
django-js-asset==3.1.2
djangorestframework==3.16.0
psycopg2==2.9.10
python-dotenv==1.1.0
requests==2.32.3
